moved common components to /components folder
