+++ src/components/Concept/ConceptPage.jsx
@@ -25,11 +25,7 @@ import Button from '@ndla/button';
 import Tabs from '@ndla/tabs';
 import { CreatedBy, Spinner } from '@ndla/ui';
 import config from '../../config';
-import {
-  ImageContent,
-  TextContent,
-  OembedContent,
-} from '../../containers/ListingPage/LicenseBox';
+import { ImageContent, TextContent, OembedContent } from '../LicenseBox';
 import { fetchArticle } from '../../api/article/articleApi';
 import { fetchConcept } from '../../api/concept/conceptApi';
 import { fetchImage } from '../../api/image/imageApi';
