+++ src/components/Listing/ListingPage.jsx
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) 2020-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import React from 'react';
+
+const ListingPage = () => {
+  return <h1>Hello world from Listing</h1>;
+};
+
+export default ListingPage;
